+++ compiler/openmp/samples/_OPENMP.c
@@ -9,6 +9,7 @@
  * - 201811：OpenMP 5.0
  * - 202011：OpenMP 5.1
  * - 202111：OpenMP 5.2
+ * - 202411: OpenMP 6.0
  */
 #include <stdio.h>
 #include <omp.h>
@@ -17,6 +18,9 @@ int main(int argc, char *argv[])
 {
 #ifdef __STDC_VERSION__
 	printf("__STDC_VERSION__ = %ld, ", __STDC_VERSION__);
+#endif
+#ifdef __OPENMP__
+	printf("__OPENMP__ = %d, ", __OPENMP__);
 #endif
 	printf("_OPENMP = %d\n", _OPENMP);
 	return 0;